move mouse control to a separate class
